refactor: devices should be a proper package
