+++ src/blinkstick/devices/__init__.py
@@ -1,0 +1,3 @@
+from .device import BlinkStickDevice
+
+__all__ = ["BlinkStickDevice"]
